Trait cfx_vm_interpreter::Finalize

pub trait Finalize {
    // Required method
    fn finalize<C: Context>(self, context: C) -> Result<FinalizationResult>;
}

Types that can be “finalized” using an EVM.

In practice, this is just used to define an inherent impl on Reult<GasLeft<'a>>.

Required Methods

fn finalize<C: Context>(self, context: C) -> Result<FinalizationResult>

Consume the context, call return if necessary, and produce call result.

Object Safety

This trait is not object safe.