Crate cfx_vm_interpreter

Re-exports

pub use self::factory::Factory;

pub use self::instructions::GasPriceTier;

pub use self::instructions::INSTRUCTIONS;

pub use self::instructions::INSTRUCTIONS;

pub use self::instructions::INSTRUCTIONS_CANCUN;

pub use self::instructions::INSTRUCTIONS_CANCUN;

Modules

factory

Evm factory.

instructions

VM Instructions list and utility functions

Macros

evm_test

Create tests by injecting different VM factories

evm_test_ignore

Create ignored tests by injecting different VM factories

Structs

FinalizationResult

Finalization result. Gas Left: either it is a known value, or it needs to be computed by processing a return instruction.

Enums

VMType

Type of EVM to use.

Traits

CostType

Cost calculation type. For low-gas usage we calculate costs using usize instead of U256

Finalize

Types that can be “finalized” using an EVM.