cfx_vm_interpreter

Trait Finalize

pub trait Finalize {
    // Required method
    fn finalize<C: Context>(self, context: C) -> Result<FinalizationResult>;
}

Types that can be “finalized” using an EVM.

In practice, this is just used to define an inherent impl on Reult<GasLeft<'a>>.

Required Methods

fn finalize<C: Context>(self, context: C) -> Result<FinalizationResult>

Consume the context, call return if necessary, and produce call result.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.